Update rc.conf
This should work to obtain an ipv6 default route on a forwarded interface.  The reason defaultroute needed to be specified previously was because forwarded interfaces do not accept rtadv from upstream.  There is a trick in linux to set accept_rtadv=2 which accepts the router information but does not attempt SLAAC.  The ipv6_cpe_wanif does the same thing on FreeBSD since 9.0.